Fixed the return value of play_sfall_sound function
* before it returned a raw pointer instead of the ID number of the
played sound.

+++ sfall/ScriptOps/MiscOps.hpp
@@ -907,7 +907,7 @@ static void _stdcall play_sfall_sound2() {
 					  &loopArg = opHandler.arg(1);
 
 	if (fileArg.isString() && loopArg.isInt()) {
-		long soundID = (long)PlaySfallSound(fileArg.strValue(), loopArg.asBool());
+		DWORD soundID = PlaySfallSound(fileArg.strValue(), loopArg.asBool());
 		opHandler.setReturn(soundID);
 	} else {
 		OpcodeInvalidArgs("play_sfall_sound");
